Update project_foto.html
After validation made with W3C, I changed the language mark in line 2, added heading to the initial section in line 14 and moved the heading out of the <ul> mark and shifted it to the line above (64). Since I checked the CSS file as well, I put the corresponding icon at the bottom of the page (lines 75-80).